refactor: scripts to test two and three hops
