+++ scripts/create_pool/config-testnet.json
@@ -1,6 +1,7 @@
 {
   "RPC_ENDPOINT_URL": "https://api-testnet.renec.foundation:8899/",
   "REDEX_PROGRAM_ID": "FJqbYpNn1JWb2qVKPZHV6emjpSuGcdk3NDYYGBxzMv8D",
+  "DISCOUNT_TOKEN": "J1ccfcMXCcQQUMS6vPquq7LaYLujuQX8AbxF5PHRKx8U",
   "PROTOCOL_FEE_RATE": 300,
   "FEE_TIERS": [
     {
