+++ scripts/versioned-tx/lookup-table.json
@@ -1,0 +1,5 @@
+{
+  "Ph4fMgWiEoVcEAkjnHGxbSaN4K2yVimXbzbSjT21cv3": "ZLvu1HVc7yDjqxL72yQxHK9wRmHxxkcdsXZUJgZkMh1",
+  "ER9g3kzTNR9EoguqjkfwqmgB6dvw8rMq3Lq8FmwaxW4e": "8GjrxBqX8AZ7J8tZhMpowwpTL5NGbA4LpRexQ6mnuzXy",
+  "FsmrjczzUv2fTFKRfFXLBkE97dDaQJwNfQm2kmw4XM9K": "EFHpuGEpb6zbSNfqDHQ39TDZhFcm9WLas4SRiRvhXDeP"
+}
